backend/requirements.txt
Flask
Flask-PyMongo
Flask-Bcrypt
python-dotenv
Flask-JWT-Extended
